+++ Parakolay_DotNet_SDK/Models/PointsResult.cs
@@ -1,0 +1,7 @@
+public class PointsResult
+{
+    public bool isSucceed { get; set; }
+    public object errorCode { get; set; }
+    public object errorMessage { get; set; }
+    public double availablePoint { get; set; }
+}