+++ CHANGELOG.md
@@ -1,6 +1,15 @@
+## 1.0.2
+
+* Installment function added.
+* Point Inquiry function added.
+
 ## 1.0.1
 
 * Reverse function added.
 * Return function added.
 * Provision return fixed, Provision function changed accordingly.
 * BIN Info function added.
+
+## 1.0.0
+
+* TODO: Initial release.